change cast instant value to float

+++ app/models/contract.rb
@@ -392,7 +392,7 @@ class Contract < ActiveRecord::Base
 
     ContractSample.compact_keys.each do |rkey|
       value = if Rails.env.production?
-                date_keys.empty? ? 0 : (($redis.hget(date_keys.last, "#{rkey[:name]}_instant").to_i / $redis.hget(date_keys.last, "total_seconds").to_i) * 8)
+                date_keys.empty? ? 0 : (($redis.hget(date_keys.last, "#{rkey[:name]}_instant").to_f / $redis.hget(date_keys.last, "total_seconds").to_f) * 8)
               else
                 rand(1024)
               end
